+++ scripts/embedding/gprolog/gprolog_embed_logtalk.sh
@@ -5,7 +5,7 @@
 ##   This script creates a new GNU Prolog top-level interpreter
 ##   that embeds Logtalk and optionally a Logtalk application
 ## 
-##   Last updated on January 9, 2024
+##   Last updated on January 30, 2025
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   SPDX-FileCopyrightText: 1998-2025 Paulo Moura <pmoura@logtalk.org>
@@ -220,12 +220,10 @@ else
 	extension=''
 fi
 
-# use GNU sed if available instead of BSD sed
-if gsed --version >/dev/null 2>&1 ; then
-	sed="gsed"
-else
-	sed="sed"
-fi
+case $(sed --help 2>&1) in
+  *GNU*) sed_i () { sed -i "$@"; };;
+  *) sed_i () { sed -i '' "$@"; };;
+esac
 
 cp "$LOGTALKHOME/adapters/gnu.pl" .
 cp "$LOGTALKHOME/core/core.pl" .
@@ -244,7 +242,7 @@ if [ "$settings" != "" ] && [ "$settings" != "none" ] ; then
 	else
 		gplgt$extension --query-goal "logtalk_compile('$settings',[optimize(on),scratch_directory('$temporary')]),halt"
 	fi
-	$sed -i "s/settings_file, allow/settings_file, deny/" gnu.pl
+	sed_i "s/settings_file, allow/settings_file, deny/" gnu.pl
 else
 	touch settings_lgt.pl
 fi