+++ scripts/embedding/sicstus/sicstus_logtalk_po.sh
@@ -6,7 +6,7 @@
 ##   compiler and runtime and optionally an application.po file with a
 ##   Logtalk application
 ## 
-##   Last updated on January 14, 2025
+##   Last updated on January 30, 2025
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   SPDX-FileCopyrightText: 1998-2025 Paulo Moura <pmoura@logtalk.org>
@@ -227,12 +227,10 @@ else
 	extension=''
 fi
 
-# use GNU sed if available instead of BSD sed
-if gsed --version >/dev/null 2>&1 ; then
-	sed="gsed"
-else
-	sed="sed"
-fi
+case $(sed --help 2>&1) in
+  *GNU*) sed_i () { sed -i "$@"; };;
+  *) sed_i () { sed -i '' "$@"; };;
+esac
 
 cp "$LOGTALKHOME/adapters/sicstus.pl" .
 cp "$LOGTALKHOME/core/core.pl" .
@@ -251,7 +249,7 @@ if [ "$settings" != "" ] && [ "$settings" != "none" ] ; then
 	else
 		sicstuslgt$extension --goal "logtalk_compile('$settings',[optimize(on),scratch_directory('$temporary')]),halt."
 	fi
-	$sed -i "s/settings_file, allow/settings_file, deny/" sicstus.pl
+	sed_i "s/settings_file, allow/settings_file, deny/" sicstus.pl
 	cat \
 		sicstus.pl \
 		paths_*.pl \