+++ scripts/embedding/xsb/xsb_logtalk_xwam.sh
@@ -6,7 +6,7 @@
 ##   and runtime and optionally an application.xwam file with a Logtalk
 ##   application
 ## 
-##   Last updated on January 14, 2025
+##   Last updated on January 30, 2025
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   SPDX-FileCopyrightText: 1998-2025 Paulo Moura <pmoura@logtalk.org>
@@ -239,12 +239,10 @@ else
 	extension=''
 fi
 
-# use GNU sed if available instead of BSD sed
-if gsed --version >/dev/null 2>&1 ; then
-	sed="gsed"
-else
-	sed="sed"
-fi
+case $(sed --help 2>&1) in
+  *GNU*) sed_i () { sed -i "$@"; };;
+  *) sed_i () { sed -i '' "$@"; };;
+esac
 
 cp "$LOGTALKHOME/adapters/xsb.pl" .
 cp "$LOGTALKHOME/core/core.pl" .
@@ -263,7 +261,7 @@ if [ "$settings" != "" ] && [ "$settings" != "none" ] ; then
 	else
 		xsblgt$extension -e "logtalk_compile('$settings',[optimize(on),scratch_directory('$temporary')]),halt."
 	fi
-	$sed -i "s/settings_file, allow/settings_file, deny/" xsb.pl
+	sed_i "s/settings_file, allow/settings_file, deny/" xsb.pl
 	cat \
 		xsb.pl \
 		paths_*.P \