+++ scripts/embedding/xvm/xvm_logtalk_pl.sh
@@ -5,7 +5,7 @@
 ##   This script creates a XVM logtalk.pl file with the Logtalk compiler and
 ##   runtime and optionally an application.pl file with a Logtalk application
 ## 
-##   Last updated on January 14, 2025
+##   Last updated on January 30, 2025
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   SPDX-FileCopyrightText: 1998-2025 Paulo Moura <pmoura@logtalk.org>
@@ -224,12 +224,10 @@ else
 	extension=''
 fi
 
-# use GNU sed if available instead of BSD sed
-if gsed --version >/dev/null 2>&1 ; then
-	sed="gsed"
-else
-	sed="sed"
-fi
+case $(sed --help 2>&1) in
+  *GNU*) sed_i () { sed -i "$@"; };;
+  *) sed_i () { sed -i '' "$@"; };;
+esac
 
 cp "$LOGTALKHOME/adapters/xvm.pl" .
 cp "$LOGTALKHOME/core/core.pl" .
@@ -248,7 +246,7 @@ if [ "$settings" != "" ] && [ "$settings" != "none" ] ; then
 	else
 		xvmlgt$extension --goal "logtalk_compile('$settings',[optimize(on),scratch_directory('$temporary')]),halt."
 	fi
-	$sed -i "s/settings_file, allow/settings_file, deny/" xvm.pl
+	sed_i "s/settings_file, allow/settings_file, deny/" xvm.pl
 	cat \
 		xvm.pl \
 		paths*.pl \